Manage case of new articles
New articles are defined in a article-definition node, hence this is the
node where the new article id should be read, and in this case the previous
text of the article is None, hence it should be created with an empty content
to apply the diff.

Also added identifiers when an error is catched to be more aware of the
buggy piece of code.