Read in text headers of diffs if the article is added or removed
Previously this was only encoded in the filename header (--- and +++),
now it is also encoded in the text header (@@), possibly I will remove
the filename header given the information about the modified article is
already known in the JSON structure. The only case where the filename
header could be useful is when the article is moved, but this information
is not provided by DuraLex/SedLex for now.