Put the amendment in context to be able to delete it
I’m not completely sure it is the right way to do it, but it works for
now. I would prefer _link_ an amendment to its related article in the
bill in order to not to mix the amendment itself and its amended text,
possibly with a special property 'context', or a global structure where
the bill would be a container for the initial text and the amendments
linked together through uuids. But all that is for later when a larger
scale will be targeted.