Encapsulate the amendment in a high-level DuraLex tree
Previously there was not really a context about the amendment and/or
the amended text; now they are encapsulated in a law-proposal node,
the bill-article (if provided) is a child of the law-proposal node,
and the amendment (if provided) is a child either of the bill-article
either the law-proposal (if the amendment adds a new article, there
is sense in wraping it in a bill-article node).

The two inputs are now strictly:
1/ an amendment, to be applied on the bill-article, except if it adds
   a new bill-article;
2/ a bill-article or an article of any in-force law, anyway this text
   is in fine always applied on the in-force law.

If no amendment is provided, the bill-article (or article of a in-force
law) is applied on the in-force law. In this case, the in-force law is
retrieved through Archéo Lex (if available).

Also, improved the error messages in the user interface.