Add 1px padding around text even at the beginning of a newline
For diffed text (red or green), this padding avoids that the letters are
sticked to the colored box. But padding for inline blocks was only for the
beginning of the inline block and was not repeated on the newlines. This is
fixed with the new CSS property+value box-decoration-break: clone.

Thanks for https://css-tricks.com/multi-line-padded-text/ and
https://developer.mozilla.org/fr/docs/Web/CSS/box-decoration-break.